requirements.txt
beautifulsoup4==4.12.3
certifi==2024.7.4
charset-normalizer==3.3.2
idna==3.8
pytaglib==3.0.0
rapidfuzz==3.9.6
requests==2.32.3
soupsieve==2.6
syncedlyrics==1.0.1
urllib3==2.2.2
